relocate store to utils
