+++ tests/solver/test_subtask.py
@@ -6,7 +6,7 @@ from inspect_ai.dataset import Sample
 from inspect_ai.model import ModelOutput
 from inspect_ai.scorer import match
 from inspect_ai.solver import Generate, TaskState, generate, solver, subtask
-from inspect_ai.solver._subtask.store import store
+from inspect_ai.util._store import store
 
 
 @subtask
